Refactore(Store): Separate Business Logic
